+++ ab-tester/src/environments/stacks_node/db.rs
@@ -1,5 +1,5 @@
-pub mod schema;
+pub mod marf;
 pub mod model;
+pub mod schema;
 pub mod stacks_burnstate_db;
 pub mod stacks_headers_db;
-pub mod marf;
